Add iso-test-2.
This is a shader experiment to draw a light around the player character
that has a sprite aesthetic. What hasn't been done is to map everything
that falls in a certain radius circle to a palette that goes from purple
to orange for colored lighting.

The shader is incomplete because I haven't figured out a way to make
this work with multiple light sources nor with the y-sorting of the
tilemap.

This uses the same assets as iso-test.