Add turn mechanic to iso-test-pkmn movement.
If the player quickly presses and lets go of a new direction, the player
character will turn but not move a square. This only happens if the
player is standing still and the direction is not the direction that the
player is already facing (at least that's how it works in Pokemon Fire
Red).

Add in a quick hack to emulate this behavior. Seems to be working 99%,
aside from some flickering because the animation is updated in a couple
different places. A good refactor should get rid of the flickers.

Also add direction to string conversion functions for less verbose code.