Re-implement hitting wall animation.
Add the animation slow down when hitting a wall and keeping the button
pressed down in the character movement controller. Add new variable to
character state "hitting_wall_animation_speed" for customizability.

Also speed up the animations to 0.15 duration instead of 0.2, which was
incorrect.

NOTE: if this character is following another character one tile behind
and both are moving at the same speed in the same direction, the walk
animation will be slow. This is incorrect, because the is_hitting_wall()
function detects a collision body in the destination tile, but this is
actually incorrect because the colliding body is moving out of the way.
Not sure how to fix this...