Create player.tscn inherited scene.
Differentiate general character scene from player scene. Right now this
is mostly just to decouple input. The character scene no longer moves or
is connected to input.

The player inherited scene has an additional script that takes input and
writes it to the character state.

This will allow for multiple sources in the future (i.e. for AI
controlled character NPCs) or something.