Test bookshelf scene instead of sprite.
Encapsulate bookshelf (and eventually all other room objects that take
up multiple tiles or are interactable) in its own scene. Trying to see
what the best way to do this is. Using tilemaps for collision is doable
and pretty good. Unfortunately, this means it doesn't align well with
the parent tilemap when y-sorting is taken into account. Maybe Godot 4
tilemap is better.

Using smaller snap step to get around this for now...