Convert all furniture to scenes.
Still kind of having trouble aligning origins to tilemap grid. This is
not a problem for entities that are only 1 tile large, but if they are 2
or more tiles, the ideal origin for y-sorting should be in the middle of
two tiles, but this makes it so that the sprite is misaligned in many
cases.

Not sure if there's a better way to do this, but it's pretty good for
now.