Create mixins class. Add interaction/interactable.
Create interaction mixin which will allow the owner of said mixin to
interact with all entities that contain an interactable mixin.

Create interactable mixin, which will act as a hook for entities
that contain this to define behavior for interaction.

Also I just realized that tilemap.world_to_map gives tile coordinates
and for some reason, the origin of each coordinate is the top middle and
not the center of the tile. Everything is working so far because the
character sprite is offset 8 pixels down so the feet end up in the
center. This is fine for the demo, but if I want to re-use this code, I
should probably change all the calculations to use the center of the
tiles.