Split input handling into its own mixin.
Interaction mixin and character movement input handling is now routed
through the input-controller mixin.

This seems unnaturally indirect, but I think it's the best way so far.
If all input goes through this single point, it will make it easier to
disable input for cases when we want to do cutscenes and dialogue.

This will also make is easier to swap out input-controller with other
things like ai-controller (for cpu controlled characters) or to make
modifications to input handling later such as for multiplayer or
something.

The con here is that you'd expect the movement or interaction mixin to
be self-contained and work if you just drag it to the entity and set the
export vars. Now, if you do that nothing will happen. You need to
include the input-controller mixin before anything will respond to
controls.