Rename a lot of stuff for no reason...
lib_game_state is now just "game". Game is now an autoloaded Scene and
not just an autoloaded script.

In the debug scene, rename the ui-layer canvaslayer to "debug-layer" and
move it to 100 instead of 1, so it can be on top of the actual ui-layer.

Rename global/scripts to global/autoload. Change the path of the
autoload scripts.

Also in this commit are stubs for dialogue-manager and spacing changes
to dialogue-box.