Solve global routing for dialogue manager.
The problem here is that I want the dialogue manager to be an object
within the scene tree but also globally available for anything to
operate dialogue. This is tough, because autoload is a way to make
things globally available, but there is no way to automatically
reference NodePaths to nodes in a scene tree from inside the autoload class.

I decided to follow a type of Service Locator pattern with a null
service default.

I think this is a little complicated, but it'll work for now.