Create DialogueData and dialogue-manager.
It's not working just yet. In this commit, you can create DialogueData
as a resource type and fill it with strings and some other settings.
Then provide it to an interactable and when you interact with it, the
interactable can call the dialogue manager to play the DialogueData.
It's not implemented correctly, so only the last string will animated.

Create DialogueData class name and make it a Resource type. This allows
me to serialize and deserailzie to tres files. I can also export it as
script member so I can use the GUI to swap out Dialogue data. This is
done for the wall-sign in the player-house-floor-2 scene.

Create dialogue-manager which is available in the game singleton for all
entities to be able to start dialogue.

The InputState enum in game has been renamed to InputMode. It just feels
more like a "mode" rather than a "state", as in state-machine.

The DialogueBoxType enum has been moved to game instead of
dialogue-manager because it's needed in DialogueData and probably a few
other places. Maybe it could go in the DialogueData definition and that
would be conceptually cleaner. The effect is the same.