Add "disable_game_input" and "interact_on_enter".
These two settings applied to DialogueData and interactable classes,
respectively, allow you to implement signs the way they are in the
original pokemon fire red.

If you walk up to a sign, it will automatically start dialogue. This is
what "interact_on_enter" does. You can also walk away or press start and
everything, so unlike talking to an NPC, this type of dialogue does not
disable the input. That is what "disable_game_input" on DialogueData is
for.

Using both of these will let you simulate sign behavior.

Also all the other small changes in this commit are all made to fix and
enable the interaction behavior of sign items.