Fix interaction bug.
If can_interact() of an interactable entity returns false, then it won't
be added to the interactables queue of character's state. If the
character leaves interactable area, the interact_event_occurred signal
will be disconnected and throw an error because it was never connected
in the first place.

This fix just avoids the check if the interactable wasn't able to be
interacted with.