Change movement mixin to use mixin hooks.
This is so the movement mixin will automatically respect the "enabled"
variable in the base class.

When you have inherited scenes, the _ready() builtin functions
automatically call .super(). Unlike how ready() is normally called
(ready is normally called bottom-up, so children are ready and parents
are only ready if all children are done), the ready() function for mixin
was called before the inherited movement class. This was causing
problems because somehow mixin was calling movement._ready() before all
exported variables and nodepaths were initalized.