Change move signals again.
Add "move" signal and make it so that "post_move" is emitted only when
player stops. This makes it more convenient for scripts that need to use
these signals to distinguish between final stop and when character moves
between tiles.

This also fixes a bug where "pre_move" is emitted twice. This was due to
the fact that coming out of a turn, we transition back into STAND before
WALK as a hack. The code has been repeated so we go directly from TURN
into WALK if the player is holding down directional keys when the turn
debounce timer times out.