Fix movement bug introduced with new TURN changes.
When transitioning from TURN to WALK state, sometimes with the right
timing, it is possible that we enter WALK state while the internal
_movement_vector variable is zero length. This was because we set the
state to WALK in the turn debounce method without setting the
destination and other necessary variables.

Not doing this causes the state machine to stall in an infinite loop
because the WALK state does not expect that _movement_vector to have
zero length.