Implement scene changing and doors.
Implement upstairs and downstairs cutscenes in code. These need to be
changed to use AnimationPlayer.

Implement scene changing interface on game autoload class. This is used
to change scenes.

Add spawner gizmo to instantiate player character in the right spot.
Note that at the game start, there is special code to spawn player
character in "new-game" spawn point in player-house-floor-2 level. This
is in main.gd but it probably should go somewhere else. (See TODO)

Create screen-wipe UI scene for fade-to-black and fade-to-normal. I can
probably refactor this to be part of some bigger cutscene framework. Or
at least get rid of all global references to it by using the
AnimationPlayer.

Fix movement-visual glitch where you can see it drawn for one frame
before it gets fully instantiated and disabled. Needed to add clause in
_draw() method to check parent debug object is_enabled().

IMPORTANT NOTES:

* upstairs and downstairs do not use PackedScene in destination export
  because the engine loads all of these at once and this causes cyclical
  crashing if they refer to each other (like floor 1 and 2 do). They
  are also all preloaded and that's not necessarily what we want.

* upstairs and downstairs should probably inherit from common entity
  called "portal" or "door" or something.

* this commit is a mess. I was learning and working out quite a few
  things. The subsequent commits should re-implement this functionality
  in a more idiomatic way (using AnimationPlayer) and possibly some
  sort of CutsceneManager framework.

* AnimationPlayer is so powerful that people actually use this to do
  (branching) dialogue too. Perhaps the DialogueManager and
  CutsceneManager can be one in the same. This will require more
  experience with the engine than I have now, so this is a long
  term goal. Probably will only come in later experiements entirely.

* Need to create custom sprite assets for spawner gizmo visualizations.
  Might also want to replace visualization for movement-visual with
  custom sprite assets.

* There needs to be some place to put the code that "initializes" each
  level once they are swapped in. The code I'm talking about is the
  call to the spawner to create the player character and the code to
  reparent the camera to the player character. Not sure where this
  should go.