Modify how interaction mixin works.
Update the interaction trigger area position based on movement mixin
signals. This is better because the volume stays constant relative to
the parent entity and is not updated every single frame.

Fall back to the old way of updating on every physics process if the
parent entity does not have a movement mixin. This shouldn't be
necessary, but keep it around just in case.