Implement commands for level switching
Another mega commit. Was doing a lot of experimenting, so things didn't
come together very cleanly.

Create Command class to execute code after the fact. Use commands to
initalize levels after they've been loaded. Since things are now
encapsulated in Command objects, they can live on the level (as with
default_load_actions) or generated and customized on the fly as levels
are switched (like in upstairs or downstairs entities).

It's a complex solution that is meant for a larger game, but it seems
very powerful. I may be able to simplify and smooth out the rough edges
over time and come up with something really cool in a few iterations for
later game ideas.