Modify debug (and entity-info) visibility toggle.
Change debug script into tool script so it runs in editor. Add code to
toggled the canvas layer visibility of the scene when the user changes
visibility (usually by clicking the eyeball of the root node).

Modifing the .tscn of entity-info was necessary for it to receive the
changes from its parent scene.