Fix movement script stairs corner case.
Change set_destination parameters so that map defaults to the _tilemap
member of the movement mixin. This makes it simpler to call and lets you
call it from animation player (where you can't specify variables in the
parameters).

Fix corner case where move signal would not be emitted if you are
against a wall, and then press turn without transitioning to STAND
state. The move signal emit function sees that
_movement_against_wall_emitted is true and does not emit movement in the
new direction. Change it so that this function checks to see if the
direction is the same and emits if not.

Small changes to upstairs/downstairs entities variable names.