Change stairs to use animations and path2d.
Cleans up the code really well. Path2d, PathFollow2D, and
RemoteTransform2D let the paths stay on the stairs entity, but the
animation code is on the character.

Use commands to perform screen-wipe and use yields to keep everything in
sync.

You will notice that upstairs-house and downstairs-house are basically
identical now. This is on purpose. They should be refactor to inherit
from a "portal" entity that does scene switching, with the relevant
items customized for the stairs.