Add PlayPortalAnimationCommand and fire-and-forget
PlayPortalAnimationCommand lets you play an animation attached to a
portal with a specified target entity. This is for showing the character
walking out of stairs on scene load.

Add a new parameter to the command class called "fire-and-forget". If
this is set to true, a multi-command will not yield until the child
command completes, but instead will start executing the command and then
immediately move on to the next. This also means that the return value
of the fire-and-forget command will not be available to subsequent
commands.