Update interaction mixin on movement enable
This is for coming into and out of cutscenes. Update the interaction
trigger area when the movement mixin is enabled or disabled.

This is to fix the case where we move the player character with an
animation inside of a cutscene and the direction is changed and then we
re-enable movement, the interaction trigger area will be in the wrong
spot until the player moves.