Implement stairs exit animations.
Christ this was hard. Got stuck for a long time figuring out why the
exit animations weren't playing before I realized it was because you
needed to disable the movement mixin inside the
PlayPortalAnimationCommand or else it would overwrite the playing
animation almost immediately.

The upstairs and downstairs entities are still mostly duplicated code,
the TODO for combining them will be done next. Switching over their code
to use the PlayPortalAnimationCommand would be good too.