Fix subtle movement bug.
Maybe the movement system is too complicated...?

Fix corner case where move signal would not be emitted if you hit a wall
and then very quickly move a tiled and ran into another wall of the same
direction.

For instance, in player-house-floor-2, go to the downstairs, and run
into the tile *next* to the entrance, so "move against wall" branch in
emit_movement_signal() is triggered. Press and hold into the wall, so
the variable is not reset. Then very quickly press "up" and "left" to
try and enter the staircase. What should happen without this fix is that
the player character is stuck against the staircase entrance because the
movement signal is not emitted.