Refactor movement animation. Add wall state.
Change the animation calls on the movement mixin so that they are called
in more sane places. Since the function is now only called if it is a
different animation, this should also make it a bit more efficient.

Add _is_moving_against_wall in character.gd so that other components can
query if this is happening to the character. This is used in
input-controller so you can still interact with objects even if you are
mashing into a wall.

Change animation player call to set the playback speed of all child
AnimationPlayers to fix a bug in playback speed that was introduced when
animations was changed to use Animation tracks.