Make all vectors stored as cartesian.
Change lib_movement functions to work exclusively with cartesian
vectors.

Change all users of lib_movement to adjust for this. Now, store all
vectors in cartesian space, and then convert to isometric direction in
character data.

This is pretty confusing and not at all explicit from the code. There's
a lot of room for improvement in future iterations of this code.