Implement Direction Flip.
Add lib_movement capability to flip a Direction enum. This is used for
quickly finding the opposite of a door's facing parameter. (Facing means
which direction the entrace of the door is. The opposite of this
direction is the direction the player must move into the door to enter
it.)

Also cancel_movement when moving into doors. This is mostly unnecessary,
except that we use this function call to reset the animation playback
speed of the AnimationPlayers and to reset _is_moving_against_wall to
false.