Remove movement and input mode disable in door animation.
This is not necessary with the new Door implementation. Also, while
convenient, disabling movement and changing input mode in this command
is probably not the way to go.

What is preferable is to use this command inside a multi-command where
the input mode and/or movement is disable and re-enabled before/after
this command.

The really important thing, is if you play multiple portal animation
commands in a row, the movement will never be re-enabled, so leaving it
like this is actually a bug.