Add keepServiceAliveAfterAppIsClosed setting.
Setting this to true will mean the service lives on even after the App
is closed. Setting this to false will mean the service stops itself
after the App is closed.

This may be useful in case you want to have game modes where the game
will track steps when the user is not playing. Conversely, users may
want the service to stop tracking steps automatically when the game is
not active.

NOTE: I ran into a snag trying to set a default value for this parameter
in the startStepCounterForegroundService() function. Apparently, Kotlin
is weird about providing default values and this compounds with JNI
weirdness so that this function's default value "disappears" and you get
godot errors saying it can't find the function if you don't supply a
parameter. Fucking stupid.